Do not always try to remove bytecode version of the test.
Trying to remove the bytecode file doesn't work if it wasn't created
in the first place.  This can happen for example under Gentoo, where
bytecode is disabled while building packages to avoid sandbox
violations when Python tries to write a bytecode file for an
already-installed module outside the package build directory.

+++ tests/QtGui/pyside_reload_test.py
@@ -23,13 +23,13 @@ def increment_module_value():
     modfile.write('Sentinel.value += 1' + os.linesep)
     modfile.flush()
     modfile.close()
-    if py3k.IS_PY3K:
-        import imp
-        cacheFile = imp.cache_from_source(dst)
-    else:
-        cacheFile = dst + 'c'
-
-    os.remove(cacheFile)
+    if not sys.dont_write_bytecode:
+        if py3k.IS_PY3K:
+            import imp
+            cacheFile = imp.cache_from_source(dst)
+        else:
+            cacheFile = dst + 'c'
+        os.remove(cacheFile)
 
 class TestModuleReloading(unittest.TestCase):
